+++ ricaun.DA4R.NUnit.Console/DA4RTestService.cs
@@ -16,6 +16,7 @@ namespace ricaun.DA4R.NUnit.Console
     public class DA4RTestService : IRunTestService
     {
         private const int MINIMAL_ENGINE_VERSION = 2018;
+        private const int TIMEOUT_MINUTES = 10;
 
         private void LogApplicationInfo()
         {
@@ -42,12 +43,20 @@ namespace ricaun.DA4R.NUnit.Console
                 if (Directory.Exists(directoryResolver))
                 {
                     var result = TestEngine.GetTestFullNames(filePath, directoryResolver);
+
+                    foreach (var ex in TestEngine.Exceptions)
+                        Log.WriteLine($"ERROR: {ex.Message}");
+
                     if (result.Length > 0) return result;
                 }
             }
             catch { }
 
             var baseTests = TestEngine.GetTestFullNames(filePath);
+
+            foreach (var ex in TestEngine.Exceptions)
+                Log.WriteLine($"ERROR: {ex.Message}");
+
             if (baseTests.Length == 0)
             {
                 Log.WriteLine($"ERROR: TestEngine.GetTestFullNames is empty, some class is breaking.");
@@ -55,14 +64,13 @@ namespace ricaun.DA4R.NUnit.Console
             return baseTests;
         }
 
-        public bool RunTests(
-            string fileToTest,
+        public bool RunTests(string fileToTest,
             int revitVersionNumber,
             Action<string> actionOutput = null,
             string forceLanguageToRevit = null,
             bool forceToOpenNewRevit = false,
-            bool forceToWaitRevit = false,
             bool forceToCloseRevit = false,
+            int timeoutMinutes = 0,
             params string[] testFilters)
         {
             LogApplicationInfo();
@@ -97,7 +105,7 @@ namespace ricaun.DA4R.NUnit.Console
             {
                 try
                 {
-                    await Run(fileToTest, revitVersionNumber, forceLanguageToRevit, actionOutput);
+                    await Run(fileToTest, revitVersionNumber, forceLanguageToRevit, timeoutMinutes, actionOutput);
                     return true;
                 }
                 catch (Exception ex)
@@ -122,13 +130,16 @@ namespace ricaun.DA4R.NUnit.Console
             return TestExceptionUtils.CreateTestAssemblyModelWithException(fileToTest, testNames, exception);
         }
 
-        private async Task Run(string filePath, int revitVersionNumber, string revitLanguage, Action<string> actionOutput)
+        private async Task Run(string filePath, int revitVersionNumber, string revitLanguage, double timeoutMinutes, Action<string> actionOutput)
         {
 
             revitLanguage = LanguageUtils.GetArgument(revitLanguage);
 
             Log.WriteLine($"Version: {revitVersionNumber}");
             Log.WriteLine($"Language: {revitLanguage}");
+
+            if (timeoutMinutes <= 0) timeoutMinutes = TIMEOUT_MINUTES;
+            if (timeoutMinutes != TIMEOUT_MINUTES) Log.WriteLine($"Timeout: {timeoutMinutes} minutes");
 
             if (!Path.GetExtension(filePath).EndsWith("dll"))
             {
@@ -150,7 +161,7 @@ namespace ricaun.DA4R.NUnit.Console
                 EnableConsoleLogger = Log.Enabled,
                 //EnableParameterConsoleLogger = true,
                 EnableReportConsoleLogger = Log.Enabled,
-                RunTimeOutMinutes = 10.0, //10.0,
+                RunTimeOutMinutes = timeoutMinutes,
                 ForgeEnvironment = App.ForgeEnvironment,
             };