+++ ricaun.DA4R.NUnit/Revit/App.cs
@@ -1,46 +1,23 @@
 ﻿using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
-using DesignAutomationFramework;
+using ricaun.DA4R.NUnit.Revit.DA;
 using ricaun.DA4R.NUnit.Services;
-using System;
 
 namespace ricaun.DA4R.NUnit.Revit
 {
-    public class App : IExternalDBApplication
+    public class App : DesignApplication
     {
-        public ExternalDBApplicationResult OnStartup(ControlledApplication application)
+        public override void OnStartup()
         {
-            Console.WriteLine("--------------------------------------------------");
-            Console.WriteLine(this.GetType().Assembly.FullName);
-            Console.WriteLine(typeof(ricaun.NUnit.TestEngine).Assembly.FullName);
-            Console.WriteLine("--------------------------------------------------");
-
-            RevitParameters.AddParameter(application, application.GetApplication());
-            DesignAutomationBridge.DesignAutomationReadyEvent += DesignAutomationBridge_DesignAutomationReadyEvent;
-            return ExternalDBApplicationResult.Succeeded;
+            
         }
-
-        public ExternalDBApplicationResult OnShutdown(ControlledApplication application)
+        public override void OnShutdown()
         {
-            DesignAutomationBridge.DesignAutomationReadyEvent -= DesignAutomationBridge_DesignAutomationReadyEvent;
-            Console.WriteLine("--------------------------------------------------");
-            return ExternalDBApplicationResult.Succeeded;
+            
         }
-
-        private void DesignAutomationBridge_DesignAutomationReadyEvent(object sender, DesignAutomationReadyEventArgs e)
+        public override bool Execute(Application application, string filePath, Document document)
         {
-            DesignAutomationBridge.DesignAutomationReadyEvent -= DesignAutomationBridge_DesignAutomationReadyEvent;
-
-            var data = e.DesignAutomationData;
-
-            Console.WriteLine("--------------------------------------------------");
-            Console.WriteLine($"RevitApp: {data.RevitApp} FilePath: {data.FilePath} RevitDoc: {data.RevitDoc}");
-            Console.WriteLine($"RevitNET: {data.RevitApp.GetType().Assembly.FullName}");
-            Console.WriteLine("--------------------------------------------------");
-
-            RevitParameters.AddParameter(data.RevitApp);
-
-            e.Succeeded = DesignAutomationController.Execute(data.RevitApp, data.FilePath, data.RevitDoc);
+            return new DesignAutomationController().Execute(application, filePath, document);
         }
     }
 }