+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] / 2024-08-20 - 2024-09-19
+### Features
+- Support bundle multiple version of `Revit`.
+### Application
+- Move `RevitParameters` to `DesignAutomationController`
+- Add `TargetFramework` in `DesignAutomationController`
+- Update `App` to use abstract `DesignApplication` to load correct assembly in the bundle.
+- Update `ricaun.NUnit` to version `1.4.0`, `CaseSource` and `ExportedTypes` support.
+- Add `ExternalServer` to force `DesignAutomationReadyEvent` event to trigger inside the server and execute with a valid `ActiveAddInId`.
+### Console
+- Update `ricaun.Autodesk.Forge.Oss.DesignAutomation` to `2.0.0`
+- Update `ricaun.Revit.Command` to version `1.5.*` to support `TestCaseSource`.
+- Update `ricaun.Revit.Command` to version `1.4.*` to support `Timeout`.
+### Tests
+- Add `Assembly` with location and `TargetFramework`.
+- Add `AddInIdTests` to make sure `ActiveAddInId` is valid.
+
 ## [1.2.0] / 2024-04-05 - 2024-08-20
 ### Application
 - Support `Revit 2025`
@@ -116,6 +133,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TestAssembly in `*.zip` and `output.json`
 
 [vNext]: ../../compare/1.0.0...HEAD
+[2.0.0]: ../../compare/1.2.0...2.0.0]
 [1.2.0]: ../../compare/1.1.4...1.2.0
 [1.1.4]: ../../compare/1.1.3...1.1.4
 [1.1.3]: ../../compare/1.1.2...1.1.3